requirements.txt
pandas
psycopg2-binary
openpyxl
python-dotenv
schedule
openai
tenacity
requests